Sending a messages to the telegram chat.
Messages to the telegram chat are sent from the "try" and "except" blocks of the main function by calling different functions: "bot_message" and "bot_message_for_except", respectively. Now errors of the "telegram.TelegramError" module are correctly handled when sending a message to a telegram chat from the "except " block of the main function.